pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "terma"
version = "0.1.0"
description = "A lightweight, Python-based CLI tool that integrates LLMs directly into your terminal"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Terma Contributors"}
]
dependencies = [
    "litellm>=1.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "tomli-w>=1.0.0",
    "rich>=13.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["terma", "defaults"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
terma = "terma.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

